[batch] Move spawnBatch function into phase module
