+++ src/batch/phase.ts
@@ -1,4 +1,10 @@
+import type { NS } from "netscript";
+
 import { CONFIG } from "batch/config";
+
+import { TAG_ARG } from "services/client/memory_tag";
+
+import { collectDependencies } from "util/dependencies";
 
 export interface BatchPhase {
     /** The script to run this phase */
@@ -58,4 +64,50 @@ export function calculatePhaseStartTimes(phases: BatchPhase[]) {
         p.start += earliest;
     }
     return phases;
+}
+
+
+/**
+ * Exec all phases in a batch on host
+ *
+ * @param ns - The NS object
+ * @param host - Host to run the batch on
+ * @param phases - Phases to exec
+ * @param donePort - Port where the last phase should send a "complete" message
+ * @returns array of pids of all phases
+ */
+export async function spawnBatch(ns: NS, host: string | null, target: string, phases: BatchPhase[], donePort: number, allocId: number): Promise<number[]> {
+    if (!host) return [];
+
+    const scripts = Array.from(new Set(phases.map(p => p.script)));
+    let dependencies = scripts.map(script => collectDependencies(ns, script)).reduce((c, s) => c.union(s));
+    ns.scp([...scripts, ...dependencies], host, "home");
+
+    let pids = [];
+    for (const [idx, phase] of phases.map((phase, idx) => [idx, phase] as [number, BatchPhase])) {
+        if (phase.threads <= 0) continue;
+        const script = phase.script;
+
+        let lastArg = idx === phases.length - 1 ? donePort : -1;
+
+        let retryCount = 0;
+        while (true) {
+            if (retryCount > CONFIG.harvestRetryMax) {
+                ns.print(`ERROR: harvest repeatedly failed to exec ${script} on ${host}`);
+                ns.ui.openTail();
+                return pids;
+            }
+
+            const pid = ns.exec(script, host, { threads: phase.threads, temporary: true }, target, phase.start, lastArg, TAG_ARG, allocId);
+            if (pid === 0) {
+                retryCount += 1;
+                ns.print(`WARN: failed to exec ${script} on ${host}, trying again with fewer threads`);
+                await ns.sleep(CONFIG.harvestRetryWait);
+            } else {
+                pids.push(pid);
+                break;
+            }
+        }
+    }
+    return pids;
 }
